Move Operation to a separate header in preparation for refactoring
